+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 * Mobile `Dialog` will scroll internally if taller than the screen.
 * FontAwesome upgraded to v6. This includes redesigns of the majority of bundled icons - please
   check your app's icon usages carefully.
-* `fmtQuantity` now handles values in the billions as well as millions.
+* `fmtQuantity` now displays values greater than one billion with `b` unit, similar to current
+  handling of millions with `m`.
 
 ### 💥 Breaking Changes